mv the adaptive_softmax_loss to tf.contrib.layers
